pyproject.toml
[project]
name = "visaire"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "manim>=0.19.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
